// private/types/mod.ts
export type DeclarationKind = "private" | "export" | "declare";

interface DocNodeBase {
  kind: DocNodeKind;
  name: string;
  location: Location;
  declarationKind: DeclarationKind;
  jsDoc?: JsDoc;
  /** All the places from where this symbol is exported.
   * Will not be set if this declaration is `declarationKind == 'export'`, and no other places are found.
   **/
  exports?: Export[];
}

export interface Export {
  name: string;
  location: Location;
}

export type DocNodeKind =
  | "moduleDoc"
  | "function"
  | "variable"
  | "enum"
  | "class"
  | "typeAlias"
  | "namespace"
  | "interface"
  | "import";

export interface DocNodeModuleDoc extends DocNodeBase {
  kind: "moduleDoc";
  jsDoc: JsDoc;
}
// ...
export interface JsDocToken {
  text: string;
  kind: 'text' | 'lineBreak' | 'link' | 'linkText';
}

export interface JsDoc {
  doc?: string;
  /** Doc-comment string split to tokens.
   **/
  docTokens?: JsDocToken[];
  tags?: JsDocTag[];
}

export type JsDocTagKind =
  | "callback"
  | "category"
  | "constructor"
  | "default"
  | "deprecated"
  | "enum"
  | "example"
  | "extends"
  | "ignore"
  | "module"
  | "param"
  | "public"
  | "private"
  | "property"
  | "protected"
  | "readonly"
  | "return"
  | "tags"
  | "template"
  | "this"
  | "typedef"
  | "type"
  | "unsupported";

export type JsDocTag =
  | JsDocTagOnly
  | JsDocTagDoc
  | JsDocTagNamed
  | JsDocTagValued
  | JsDocTagTyped
  | JsDocTagNamedTyped
  | JsDocTagParam
  | JsDocTagReturn
  | JsDocTagTags
  | JsDocTagUnsupported;

export interface JsDocTagBase {
  kind: JsDocTagKind;
}

export interface JsDocTagOnly extends JsDocTagBase {
  kind:
    | "constructor"
    | "ignore"
    | "module"
    | "public"
    | "private"
    | "protected"
    | "readonly";
}

export interface JsDocTagDoc extends JsDocTagBase {
  kind: "category" | "deprecated" | "example";
  doc?: string;
  docTokens?: JsDocToken[];
}

export interface JsDocTagNamed extends JsDocTagBase {
  kind: "callback" | "template";
  name: string;
  doc?: string;
  docTokens?: JsDocToken[];
  tsType?: TsTypeDef;
  typeParams?: TsTypeParamDef[];
}

export interface JsDocTagValued extends JsDocTagBase {
  kind: "default";
  value: string;
  doc?: string;
  docTokens?: JsDocToken[];
}

export interface JsDocTagTyped extends JsDocTagBase {
  kind: "enum" | "extends" | "this" | "type";
  type: string;
  tsType?: TsTypeDef;
  doc?: string;
  docTokens?: JsDocToken[];
}

export interface JsDocTagNamedTyped extends JsDocTagBase {
  kind: "property" | "typedef";
  name: string;
  type: string;
  doc?: string;
  docTokens?: JsDocToken[];
}

export interface JsDocTagParam extends JsDocTagBase {
  kind: "param";
  name: string;
  type?: string;
  tsType?: TsTypeDef;
  optional?: true;
  default?: string;
  doc?: string;
  docTokens?: JsDocToken[];
}

export interface JsDocTagReturn extends JsDocTagBase {
  kind: "return";
  type?: string;
  doc?: string;
  docTokens?: JsDocToken[];
}

export interface JsDocTagTags extends JsDocTagBase {
  kind: "tags";
  tags: string[];
}

export interface JsDocTagUnsupported extends JsDocTagBase {
  kind: "unsupported";
  value: string;
}
// ...
export interface Location {
  filename: string;
  line: number;
  col: number;
  /** If the filename corresponds to one of the entry points passed to `createDenoProgram()`.
   * `filename` here is always absolute URL, even if a relative path was passed.
   **/
  entryPointNumber?: number;
}